Separate text UI code out into a separate module
Now it's separate, whenever we encounter a problem, we prompt to user to
automatically fix it, show a diff of what we'll change, or if we don't
support automatically fixing it, give the user the chance to edit the
config themselves and then retry.

Retrying in the case of allow_unsafe isn't working yet.