Ignore API use on A->B where X->A would classify same API use
The primary motiviation is that it allows us to avoid misclassifying a
generic function as using an API if it only uses that API via an
associated type on a generic parameter to that function.

It does mean that such a function can now use that API not via generics
too, but that doesn't seem like such a problem, since any reference to
the function with those generics would be flagged as using the API.
Also, if there are any references to the function with different
generics that don't trigger that API, then the API usage would be
flagged in the function.