Remove some unnecessary state
Instead of tracking whether a crate's config has been used, we just
check against the crate index. This fixes a bug where the state could be
wrong if the config for a crate was added via the UI, it would then warn
about the crate's config being unused. Rather than fixing it so that the
state got updated at the correct time, we just removed the state
entirely.