Rename build script before we notify parent process of linking
Previously we were running the linker, send RPC then rename the build
script. Now we run the linker, rename the build script then send the
RPC. Because of the change of order, we now put the renamed build script
as the output file in the RPC.

The motivation for this change is that we want RPCs to be able to be
replyed. As it was before, replay wouldn't work because once the rename
happened the cackle binary would be at the path pointed to by the RPC.