optimisation: Allow Names to be borrowed
This does give a small speedup, but not much, however I think it opens
opportunities for further optimisations, so I'll move forward with it.

It also introduces a bit of duplication. The code in utf8.rs is mostly
the same as the code in bytes.rs. Creating Names from mangled symbols
and non-mangled strings now don't share code, when they really should.
Both of these issues will be addressed in subsequent commits.