refactoring: Unify name splitting
We had two ways to split names, one for mangled symbols and one for
non-mangled strings. This commit adds an iterator for non-mangled
strings that is compatible with our demangling iterator. We then use
this to unify the code that splits names.

Also made use of the non-mangled iterator to test for consistency with
rustc-demangle.