Use debug information to determine from-symbol
This is necessary when functions get inlined. For example if a -> b -> c
where `c` is a restricted API and `b` is inlined into `a`, then we want
to attribute that API usage to `b` not `a`.

This increases cackle's self-analysis time from 1.343 to 1.643. It's
likely that there are ways we can recover some of this time back later
though.