Remove support for grouping API usages by crate
The only place this was used was by the null-UI, but it was adding
complexity all over the place. If we want to group multiple API usages
by crate in the null-UI, we should just implement it there. For now, we
just drop it though, since it's not obvious that grouping by crate is
all that important.