Fix a false positive attribution
This was a pattern observed in the tracing/tracing-core crates that
caused problems if tracing macros were called from a restricted API in a
higher level crate such as hyper. If all of hyper was classified as the
"net" API, then we were incorrectly classifying tracing-core as using
the "net" API because an inline(always) function from tracing-core was
referencing a variable defined by a macro in tracing that was invoked
from hyper.

This commit adds a similar, but minimal repro to the test crates and
fixes it by ignoring off-tree usages where the outer (non-inlined)
function is from the same crate as the restricted API being accessed.

Fixes #7