Allow temporary directory to be provided externally
This is useful for testing since if we let cackle generate a temporary
directory then it changes with every run, which can cause many of our
test crates to get rebuilt due to them depending on CACKLE_SOCKET_PATH,
which changes when the tmpdir changes. This gives us the control
required to have these tests not need to be rebuilt.